aubuf: the ajb mode should respect wish size
If `audio_buffer_type` is set to `adaptive` and the network does not jitter
noticeably but there might be very sporadic late audio frames then the wish
size can be increased by the user. What this commit does is to keep the average
buffer size between boundaries that respect the given wish size.