aubuf: allow zero as min_sz/wish_sz
The aubuf collects data until wish size is reached before returning real data
when the read function is called. A wish size of zero means, that data is
returned as soon there was pushed some.

On first read old data is dropped in order to reduce the latency. This is done
only if min_sz (wish_sz internally) was set to a positive value.