aubuf: prevent faulty timestamps (#75)
* aubuf: prevent faulty timestamps for partial read

There are applications where a useful timestamp is not available. E.g. if audio
files are read and decoded. Then audio frames are written to aubuf with
`timestamp == 0`. In such cases a partial read with `aubuf_read_auframe()` set
a wrong timestamp to the rest frame.

* aubuf: set correct timestamps if app provides zero

For partial read timestamps are changed for the rest frames. This works only
if the timestamps are already set correctly during writing.

This commit handles also the case correct if the application provides
timestamps and the first frame has timestamp zero and the first read is
partial.

* aubuf,auframe: move auframe_bytes_to_timestamp() to auframe

* aubuf: reset also written_sz in aubuf_flush()